requirements.txt
Django
South
factory_boy
django-debug-toolbar
webtest
django_webtest
psycopg2
gunicorn
docopt
Pillow
django-widget-tweaks
djangorestframework
robotframework
robotframework-selenium2library

